fix: update vendor dashboard template loader and improve select2 field styling
