+++ includes/DependencyManagement/Providers/CommonServiceProvider.php
@@ -3,6 +3,7 @@
 namespace WeDevs\Dokan\DependencyManagement\Providers;
 
 use WeDevs\Dokan\DependencyManagement\BaseServiceProvider;
+use WeDevs\Dokan\Shortcodes\FullWidthVendorLayout;
 
 class CommonServiceProvider extends BaseServiceProvider {
     /**
@@ -24,6 +25,7 @@ class CommonServiceProvider extends BaseServiceProvider {
         \WeDevs\Dokan\Commission\RecalculateCommissions::class,
         \WeDevs\Dokan\Order\RefundHandler::class,
         \WeDevs\Dokan\Exceptions\Handler::class,
+        FullWidthVendorLayout::class
 	];
 
 	/**
